Move things around in address
